+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ---------------
 
 * Changed keyword-only arguments without defaults to required flags
+* Added support for all variants of ``param`` and ``type``
 
 2.0.1 (2016-09-13)
 ------------------